// src/app/demo/api/cours.ts
export interface Cours {
  _id?: string; // Optional if it's generated by MongoDB
  title?: string;
  description?: string;
  formation?: string ; // Assuming it can be a string ID or a Formation object
  teacher?: string; // Assuming teacher ID is a string
  videoUrl?: string; // Assuming videoUrl will store the URL of the video
}